Add files via upload
This code corresponds to the FAS app, which has been created using the Java programming language and the Android Studio IDE. Employing a combination of a fingerprint sensor and a database, the FAS app streamlines and enhances the attendance process, ensuring efficiency and convenience. Applicable in educational institutions such as schools and universities, as well as in office environments, the Fingerprint Attendance System (FAS) facilitates the smooth monitoring of attendance for both students and employees.